refactor: rename socket to net mod
